requirements.txt
amqp==2.6.1
backports-abc==0.5
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
Django==1.11.29
django-celery-results==1.2.1
flower==0.9.5
futures==3.3.0
humanize==1.0.0
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
multi-key-dict==2.0.3
pathlib2==2.3.5
pbr==5.4.5
prometheus-client==0.8.0
PyMySQL==0.10.0
python-jenkins==1.7.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
scandir==1.10.0
singledispatch==3.4.0.3
six==1.15.0
tornado==5.1.1
urllib3==1.25.10
vine==1.3.0
zipp==1.2.0
